Move all factorization related code from util.py to a dedicated module
